+++ CHANGES.md
@@ -11,6 +11,13 @@
 
 ## 2025.4
 
+- [CHANGE] サイマルキャストの設定値である SimulcastRid を SimulcastRequestRid に移行する
+  - これによりシステム条件の WebRTC SFU Sora のバージョンが 2025.2.0 以降になる
+  - @zztkm
+- [UPDATE] システム条件を更新する
+  - WebRTC SFU Sora 2025.2.0 以降
+  - @zztkm
+
 ### misc
 
 - [UPDATE] `Claude Assistant` を `shiguredo/claude-code-action` を使うように更新する